Remove `ReplyCommentEditor` into its own class
